Update Courses.php
Hi everyone,

I'm proposing a change to the codebase to replace all instances of ' 'save'  with  _save ' This change is necessary to avoid errors in translation.

Currently, the translations for many languages, including Spanish, contain errors. This is because the translation strings are not properly formatted. By replacing ' 'save'  with  _save ' we can ensure that the translation strings are properly formatted and that the translations are accurate.

I have already created a new Spanish translation called "Spanish Venezuela" that is correctly translated and does not contain any errors. However, this translation has not yet been accepted.

In addition to the change to "'save'", there are several other documents where the same change needs to be made. I have listed these documents below:

Document 1: TransferredOutModal.php
Document 2: URL systempreference.php

I would appreciate it if you could review these changes and let me know if you have any questions.

Thank you,